Moved TeleBot to telebot_class.py Refactored TeleBot to implement apihelper.TelegramApiInterface
Changes to the API should now only reflect in the types.py and apihelper.py files, and the TeleBot class should no longer be affected each time a change is made to the API.

Multithreading functionality is removed from TeleBot.
