+++ examples/detailed_example/detailed_example.py
@@ -1,3 +1,5 @@
+from __future__ import print_function
+
 """
 This is a detailed example using almost every command of the API
 """
@@ -33,7 +35,7 @@ def get_user_step(uid):
     else:
         knownUsers.append(uid)
         userStep[uid] = 0
-        print "New user detected, who hasn't used \"/start\" yet"
+        print("New user detected, who hasn't used \"/start\" yet")
         return 0
 
 
@@ -45,11 +47,11 @@ def listener(messages):
     for m in messages:
         if m.content_type == 'text':
             # print the sent message to the console
-            print str(m.chat.first_name) + " [" + str(m.chat.id) + "]: " + m.text
+            print(str(m.chat.first_name) + " [" + str(m.chat.id) + "]: " + m.text)
 
 
 bot = telebot.TeleBot(TOKEN)
-bot.set_update_listener(listener)  # register listener
+bot.add_update_listener(listener)  # register listener
 
 
 # handle the "/start" command
